Change id type from long to int

+++ eventPlanner/src/main/java/com/ftn/iss/eventPlanner/controller/CommentController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/comments")
 public class CommentController {
     @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<UpdatedCommentDTO> updateComment(@RequestBody UpdateCommentDTO comment, @PathVariable Long id)
+    public ResponseEntity<UpdatedCommentDTO> updateComment(@RequestBody UpdateCommentDTO comment, @PathVariable int id)
             throws Exception {
         UpdatedCommentDTO updatedComment = new UpdatedCommentDTO();
 
@@ -23,7 +23,7 @@ public class CommentController {
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<?> deleteComment(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteComment(@PathVariable("id") int id) {
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
